refactor: move around codegen modules to clear out domain boundaries
